added reconstructor which adds SessionStartEvent and SessionEndEvent based on scanning all records in a log. Requires the addition of an EmptyRecord in the log. Also fixed package structure and removed more illegal logger calls.
