+++ analysis/src/main/java/org/iobserve/analysis/ServiceConfiguration.java
@@ -26,10 +26,10 @@ import org.iobserve.analysis.configurations.MultiInputObservationConfiguration;
 import org.iobserve.analysis.service.updater.AllocationVisualizationStage;
 import org.iobserve.analysis.service.updater.DeploymentVisualizationStage;
 import org.iobserve.analysis.service.updater.UndeploymentVisualizationStage;
-import org.iobserve.analysis.snapshot.SnapshotBuilder;
 import org.iobserve.analysis.toggle.FeatureToggle;
 import org.iobserve.model.correspondence.ICorrespondence;
 import org.iobserve.model.provider.neo4j.IModelProvider;
+import org.iobserve.model.snapshot.SnapshotBuilder;
 import org.iobserve.stages.general.ConfigurationException;
 import org.palladiosimulator.pcm.allocation.Allocation;
 import org.palladiosimulator.pcm.core.composition.AssemblyContext;
