+++ analysis/src/main/java/org/iobserve/analysis/configurations/AbstractObservationConfiguration.java
@@ -35,7 +35,6 @@ import org.iobserve.analysis.deployment.NetworkLink;
 import org.iobserve.analysis.deployment.UndeploymentCompositeStage;
 import org.iobserve.analysis.session.CollectUserSessionsFilter;
 import org.iobserve.analysis.session.SessionAcceptanceFilter;
-import org.iobserve.analysis.snapshot.SnapshotBuilder;
 import org.iobserve.analysis.systems.jpetstore.JPetStoreCallTraceMatcher;
 import org.iobserve.analysis.systems.jpetstore.JPetStoreTraceAcceptanceMatcher;
 import org.iobserve.analysis.systems.jpetstore.JPetStoreTraceSignatureCleanupRewriter;
@@ -45,6 +44,7 @@ import org.iobserve.analysis.traces.TraceOperationCleanupFilter;
 import org.iobserve.analysis.traces.TraceReconstructionCompositeStage;
 import org.iobserve.model.correspondence.ICorrespondence;
 import org.iobserve.model.provider.neo4j.IModelProvider;
+import org.iobserve.model.snapshot.SnapshotBuilder;
 import org.iobserve.stages.general.EntryCallStage;
 import org.iobserve.stages.general.IEntryCallTraceMatcher;
 import org.iobserve.stages.general.RecordSwitch;
