+++ analysis/src/main/java/org/iobserve/analysis/configurations/MultiInputObservationConfiguration.java
@@ -19,10 +19,10 @@ import kieker.common.configuration.Configuration;
 
 import org.iobserve.analysis.clustering.EAggregationType;
 import org.iobserve.analysis.clustering.EOutputMode;
-import org.iobserve.analysis.snapshot.SnapshotBuilder;
 import org.iobserve.analysis.toggle.FeatureToggle;
 import org.iobserve.model.correspondence.ICorrespondence;
 import org.iobserve.model.provider.neo4j.IModelProvider;
+import org.iobserve.model.snapshot.SnapshotBuilder;
 import org.iobserve.service.source.MultipleConnectionTcpCompositeStage;
 import org.iobserve.stages.general.ConfigurationException;
 import org.palladiosimulator.pcm.allocation.Allocation;
