+++ analysis/src/main/java/org/iobserve/analysis/service/generation/ModelModificationFactory.java
@@ -22,8 +22,8 @@ import java.util.List;
 import org.apache.commons.cli.CommandLine;
 import org.codehaus.plexus.personality.plexus.lifecycle.phase.InitializationException;
 import org.eclipse.emf.common.util.URI;
-import org.iobserve.analysis.snapshot.SnapshotBuilder;
 import org.iobserve.model.PCMModelHandler;
+import org.iobserve.model.snapshot.SnapshotBuilder;
 import org.palladiosimulator.pcm.allocation.Allocation;
 import org.palladiosimulator.pcm.core.composition.AssemblyContext;
 import org.palladiosimulator.pcm.repository.Repository;
