+++ collector/build.gradle
@@ -7,7 +7,7 @@ dependencies {
 	clover 'org.openclover:clover:4.2.0'
 	
 	compile project(":service")
-	compile project(":analysis")
+	compile project(":stages")
 
 	testCompile "junit:junit:4.12"
 }
