+++ collector/collector.properties
@@ -41,7 +41,7 @@ org.iobserve.service.source.MultipleConnectionTcpCompositeStage.capacity=8192
 # the data will be written to your temp folder
 
 #####
-#kieker.monitoring.writer=kieker.monitoring.writer.FileWriter
+kieker.monitoring.writer=kieker.monitoring.writer.filesystem.FileWriter
 #
 ## In order to use a custom directory, set customStoragePath as desired. 
 ## Examples: /var/kieker or C:\\KiekerData (ensure the folder exists).
@@ -87,11 +87,11 @@ kieker.monitoring.writer.filesystem.FileWriter.logStreamHandler=kieker.monitorin
 #
 ## When flushing is disabled, it could require a lot of records before
 ## finally any writing to the log file(s) is done.
-kieker.monitoring.writer.filesystem.FileWriter.flush=false
+kieker.monitoring.writer.filesystem.FileWriter.flush=true
 #
 ## When flushing is disabled, records are buffered in memory before written.
 ## This setting configures the size of the used buffer in bytes.
-kieker.monitoring.writer.filesystem.BinaryFileWriter.bufferSize=8192
+kieker.monitoring.writer.filesystem.FileWriter.bufferSize=8192
 #
 ## When compression is enabled, each log file is written as zipped binary file.
 # Please note compression results in slower record processing.
