+++ model/src/main/java/org/iobserve/model/protocom/PcmMapping.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
-package org.iobserve.analysis.protocom;
+package org.iobserve.model.protocom;
 
 import java.util.ArrayList;
 import java.util.List;
