+++ model/src/main/java/org/iobserve/model/snapshot/package-info.java
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  ***************************************************************************/
-package org.iobserve.analysis.snapshot;
+package org.iobserve.model.snapshot;
