+++ planning/src/main/java/org/iobserve/planning/ModelTransformer.java
@@ -26,13 +26,13 @@ import de.uka.ipd.sdq.pcm.designdecision.DecisionSpace;
 import org.codehaus.plexus.personality.plexus.lifecycle.phase.InitializationException;
 import org.eclipse.emf.common.util.URI;
 import org.eclipse.emf.ecore.util.EcoreUtil;
-import org.iobserve.analysis.snapshot.SnapshotBuilder;
 import org.iobserve.model.PCMModelHandler;
 import org.iobserve.model.factory.DesignDecisionModelFactory;
 import org.iobserve.model.provider.file.AllocationModelHandler;
 import org.iobserve.model.provider.file.CostModelHandler;
 import org.iobserve.model.provider.file.DesignDecisionModelHandler;
 import org.iobserve.model.provider.file.ResourceEnvironmentModelHandler;
+import org.iobserve.model.snapshot.SnapshotBuilder;
 import org.iobserve.planning.data.AllocationGroup;
 import org.iobserve.planning.data.AllocationGroupsContainer;
 import org.iobserve.planning.data.PlanningData;
