+++ rac.creator/src/main/java/org/iobserve/rac/creator/filter/PcmCorrespondentMethodStage.java
@@ -23,7 +23,7 @@ import kieker.common.record.flow.trace.operation.AbstractOperationEvent;
 import teetime.framework.AbstractConsumerStage;
 import teetime.framework.OutputPort;
 
-import org.iobserve.analysis.protocom.PcmCorrespondentMethod;
+import org.iobserve.model.protocom.PcmCorrespondentMethod;
 import org.iobserve.rac.creator.data.ClassAndMethod;
 
 /**
