+++ reconstructor/reconstructor.properties
@@ -41,7 +41,7 @@ org.iobserve.service.source.FileSourceCompositeStage.sourceDirectories=/home/rei
 # the data will be written to your temp folder
 
 #####
-#kieker.monitoring.writer=kieker.monitoring.writer.FileWriter
+kieker.monitoring.writer=kieker.monitoring.writer.filesystem.FileWriter
 #
 ## In order to use a custom directory, set customStoragePath as desired. 
 ## Examples: /var/kieker or C:\\KiekerData (ensure the folder exists).
@@ -91,7 +91,7 @@ kieker.monitoring.writer.filesystem.FileWriter.flush=true
 #
 ## When flushing is disabled, records are buffered in memory before written.
 ## This setting configures the size of the used buffer in bytes.
-kieker.monitoring.writer.filesystem.BinaryFileWriter.bufferSize=8192
+kieker.monitoring.writer.filesystem.FileWriter.bufferSize=8192
 #
 ## When compression is enabled, each log file is written as zipped binary file.
 # Please note compression results in slower record processing.
