+++ service.privacy.violation/src/main/java/org/iobserve/service/privacy/violation/data/PCMEntryCallEvent.java
@@ -15,7 +15,7 @@
  ***************************************************************************/
 package org.iobserve.service.privacy.violation.data;
 
-import org.iobserve.analysis.protocom.PcmEntity;
+import org.iobserve.model.protocom.PcmEntity;
 
 /**
  * @author Reiner Jung
